Updated code to get down to 138 cycles
Found out how to get pretty close to the 136 cycle minimum. Changes are that instead of a singular loop or multiple exas one just takes a "fast lane" in which the bulk of the work of large files is created on the fly. Additionally a medium fast cycle has been implemented to further optimize the cycles. The low speed cycle is just adding the final numbers. Further optimization efforts have to be tested with prime numbers, or numbers dividing 99 (the largest number in this level), etc.